+++ ExampleApp/src/main/java/im/dino/dbview/exampleapp/MainActivity.java
@@ -19,6 +19,11 @@ public class MainActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
+    }
+
+    @Override
+    protected void onStart() {
+        super.onStart();
         createData();
     }